updated csv reset

+++ matcher/output.py
@@ -1,4 +1,5 @@
 import os
+import sys
 from time import time
 from datetime import datetime
 from csv import DictWriter
@@ -190,16 +191,20 @@ def update_csv_betfair(
 
 
 def reset_csv():
+    if not os.path.isdir("stats"):
+        os.mkdir("stats")
     now = datetime.now().strftime("%d-%m-%Y")
     RETURNS_HEADER = "date_of_race,horse_name,bookie_odds,venue,ew_stake,balance,rating,expected_value,expected_return,win_stake,place_stake,win_odds,place_odds,betfair_balance,max_profit,is_lay,win_matched,lay_matched,arbritrage_profit,place_payout,balance_in_betfair,current_time"
     RETURNS_BAK = os.path.join(BASEDIR, "stats/returns-%s.csv" % now)
 
-    create_new_returns = input(
-        "Create new return.csv? (Recommended for new user) [Y]/n "
-    ).lower()
+    if os.path.isfile(RETURNS_BAK):
+        create_new_returns = input(
+            "Create new return.csv? (Recommended for new user) [Y]/n "
+        ).lower()
 
-    if create_new_returns != "n":
-        os.rename(RETURNS_CSV, RETURNS_BAK)
+    if create_new_returns != "n" or not os.path.isfile(RETURNS_BAK):
+        if create_new_returns != "n":
+            os.rename(RETURNS_CSV, RETURNS_BAK)
         with open(RETURNS_CSV, "w", newline="") as returns_csv:
             returns_csv.write(RETURNS_HEADER)
             print("Created new returns.csv")
